backend/requirements.txt
anyio==3.6.1
async-timeout==4.0.2
click==8.1.3
colorama==0.4.5
Deprecated==1.2.13
fastapi==0.79.1
h11==0.13.0
httptools==0.4.0
idna==3.3
mongoengine==0.24.2
numpy==1.23.2
packaging==21.3
pandas==1.4.3
pydantic==1.9.2
pymongo==4.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
rq==1.11.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
watchfiles==0.16.1
websockets==10.3
wrapt==1.14.1
requests~=2.28.1